Updated imagetocsv to be a command line tool
